+++ STM32F103C8/Src/filter.h
@@ -1,0 +1,11 @@
+//#pragma once
+
+#ifndef __FILTER_H
+#define __FILTER_H
+
+float dFilter_1(float invar);
+float dFilter_2(float invar);
+float dFilter_3(float invar);
+float dFilter_4(float invar);
+
+#endif