+++ CHANGELOG.md
@@ -5,6 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.1.0] - 2018-11-09
+### Added
+- Determinate sample rate ADC conversion
+- Filter ADC data
+- DMA for ADC
+### Fixed
+- Some mistakes
+### Removed
+- ADC Injected conversion mode
+- Platform Support STM32F407VG (Debug. Not used) 
+
 ## [0.0.1] - 2018-01-25
 ### Added
 - Basic functionality